fix logging, add responseCode enum
added a bunch of logic about handling the transactionResponse. I think
branching on the responseCode is a good way to go, because 'Approved'
should reliably indicate a transfer of funds. all other branches are
payment.fail()

moved messages calls to inside the for loops, so multiple errors will be
shown to user.