add state selector and cast productionEnabled
the state selector is quite primitive, not sure it was necessary. I had
to zip the list to itself because it wants seperate values for storage
and display.

productionEnabled wasn't working because it was deserializing as a
string. I looked in the sqlite database and found the key in
table pretixbase_event_settings. I'm not sure how type is indiciated by
django-hierarkey, but every field and a column with '1'. I just used the
get(as_type=bool) param to cast it.